Start work on adding glium render backend
