+++ src/render/gfx/asset_load.rs
