+++ src/render/gfx/buffer_util.rs
